requirements.txt
PySimpleGUI==4.47.0
speedtest-cli==2.1.3
